analytics/src/requirements.txt
grpcio
grpcio-tools
flask
nltk
